requirements.txt
torch
torchvision
opencv-python
pillow
matplotlib
tqdm
scipy
pytorch-lightning
pytorch-msssim